move get_vccs_client and improve it's usability in tests
